requirements.txt
# Python 3.12 兼容依赖
# 移除所有conda相关依赖，使用纯pip包

# 核心机器学习
torch>=2.0.0
numpy>=1.24.0
gymnasium>=0.28.0  # 替代gym，支持Python3.12

# X11截屏和事件
python-xlib>=0.33
Pillow>=9.5.0
opencv-python>=4.8.0

# 数据处理和可视化
matplotlib>=3.7.0
tqdm>=4.65.0
tensorboard>=2.13.0

# 系统工具
psutil>=5.9.0
pynput>=1.7.6  # 键盘鼠标事件

# 开发工具
pytest>=7.4.0
black>=23.0.0
